+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/HandledErrorOverrideScenario.swift
@@ -22,7 +22,8 @@ class HandledErrorOverrideScenario: Scenario {
         Bugsnag.notifyError(error) { report in
             report.errorMessage = "Foo"
             report.errorClass = "Bar"
-            report.depth += 2
+            var depth: Int = report.value(forKey: "depth") as! Int
+            report.setValue(depth + 2, forKey: "depth")
             report.metadata["account"] = [
                 "items": [400,200]
             ]